fix: cortex defaulting
The defaulting step was concatenating the list of members in the
memberlist gossip ring each time the CRD resource was updated. This
was a byproduct of the defaulting step mutating the config in the
resource.

The create and update validation webhooks now merge the user config
with the opinionated defaults.

Signed-off-by: Simão Reis <sreis@opstrace.com>